+++ arrow-udf/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix returning list in struct type. e.g. `struct<a: list<int>>` is now supported.
+
 ## [0.1.0] - 2024-01-13
 
 ### Added